requirements.txt
fhir.resources
pymongo
fastapi
uvicorn
sqlalchemy
pydantic
gunicorn